% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT.R
\name{as.data.frame.maxT}
\alias{as.data.frame.maxT}
\title{as.data.frame.maxT}
\usage{
\method{as.data.frame}{maxT}(x, ...)
}
\arguments{
\item{x}{\linkS4class{maxT}}

\item{...}{..}
}
\description{
as.data.frame.maxT
}
