% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELISpot.R
\name{as.ELISpot}
\alias{as.ELISpot}
\title{Create \linkS4class{ELISpot} Object}
\usage{
as.ELISpot(formula, data, pattern0, pattern1, ...)
}
\arguments{
\item{formula}{\link[stats]{formula}}

\item{data}{\link[base]{data.frame}}

\item{pattern0}{\link[base]{character} scalar of name of negative-control arm, or of \link[base]{regex} control pattern}

\item{pattern1}{\link[base]{regex}}

\item{...}{potential parameters, currently not in use}
}
\value{
Functions \code{\link[=as.ELISpot]{as.ELISpot()}} both return an \linkS4class{ELISpot} object
}
\description{
..
}
\keyword{internal}
