% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combn.R
\name{combn_}
\alias{combn_}
\title{Permuted Indices for Treatment}
\usage{
combn_(x)
}
\arguments{
\item{x}{see \strong{Usage}}
}
\value{
The \code{S3} generic function \code{\link[=combn_]{combn_()}} returns
a \link[base]{list} of
\link[base]{integer} \link[base]{vector}s,
as returned by the function \link[utils]{combn}.
}
\description{
Permuted indices for treatment.
}
\note{
Westfall & Young's \linkS4class{maxT} algorithm
(Box 2, page 82 of \doi{10.1214/ss/1056397487})
used the term \emph{permutation}.

However, according to \strong{R} nomenclature,
this is a \emph{combination}
(via the function \link[utils]{combn}).

The author uses the function name \code{\link[=combn_]{combn_()}} because,
unfortunately, the function \link[utils]{combn} is
not an \code{S3} generic function.
}
\keyword{internal}
