% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combn_elispot.R
\name{combn_ELISpot}
\alias{combn_ELISpot}
\title{Permutation Indices of Treatment in \code{ELISpot}}
\usage{
combn_ELISpot(data)
}
\arguments{
\item{data}{a \code{ELISpot} object}
}
\value{
Function \code{\link[=combn_ELISpot]{combn_ELISpot()}} returns a \link[base]{list} of \link[base]{integer} \link[base]{vector}s.
}
\description{
Permuted indices of treatment in a \code{ELISpot} object
}
\note{
Westfall & Young's \linkS4class{maxT} algorithm
(Box 2, page 82 of \doi{10.1214/ss/1056397487}) says \emph{permutation}.

In \strong{R} convention, this is a \emph{combination} \link[utils]{combn}.
}
\keyword{internal}
