% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFR.R
\name{DFR}
\alias{DFR}
\title{Distribution Free Resampling}
\usage{
DFR(data, ...)
}
\arguments{
\item{data}{an \code{elispot}}

\item{...}{additional parameters for function \link{split.elispot} and \link{maxT_moodie}}
}
\description{
..
}
\examples{
(m0 = moodie2ELISpot(`^a[1-9]$` ~ antigen | day + id, data = moodie, control = 'negctl')) # no log
(r1a = DFR(m0, null.value = 0))
(r2a = DFR(log(m0, base = 10), null.value = log10(2), resample = 'boot'))

}
\references{
Original algorithm by Zoe Moodie on March 16, 2010, and
her 2006 papers \doi{10.1016/j.jim.2006.07.015},
and an empirical study \doi{10.1007/s00262-010-0875-4}.

\url{https://rundfr.fredhutch.org}
}
