% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELISpot.R
\name{elispot_S3}
\alias{elispot_S3}
\alias{log.elispot}
\alias{split.elispot}
\title{Several S3 method dispatches for \code{elispot}}
\usage{
\method{log}{elispot}(x, base = exp(1))

\method{split}{elispot}(x, f = eval(call("~", attr(x, which = "design")[[3L]])), ...)
}
\arguments{
\item{x}{an \code{elispot}}

\item{base}{\link[base]{numeric} scalar, see \link[base]{log}}

\item{f}{see \link[base]{split.data.frame}, default}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{log.elispot} returns an \code{elispot}.

Function \link{split.elispot} returns a \link[base]{list} of \code{elispot}s.
}
\description{
\code{elispot} dispatches for S3 generics \link[base]{log}, \link[base]{split}.
}
\note{
Unfortunately \link[base]{log10} and \link[base]{log1p} are not an S3 generics.

\link[base]{log1p} does not have \code{base} parameter.
}
\examples{
(zoe1 = log(moodie2ELISpot(`^a[1-9]$` ~ antigen | day + id, data = moodie, control = 'negctl')))
split(zoe1)
}
