% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_d.R
\docType{class}
\name{free_d}
\alias{free_d}
\alias{free_d-class}
\alias{free_d.ELISpot}
\title{Distribution \linkS4class{free_d}-Statistic}
\usage{
free_d(x, ...)

\method{free_d}{ELISpot}(x, pid1, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{additional parameters, currently not in use}

\item{pid1}{(optional) \link[base]{integer} \link[base]{vector},
permuted indices for treatment (via the function \code{\link[=permID]{permID()}})}
}
\description{
Distribution \linkS4class{free_d} statistic to compare a treatment against a control.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\link[base]{numeric} \link[base]{vector}, distribution-free \eqn{d}-statistic}

\item{\code{null.value}}{\link[base]{numeric} scalar \eqn{c}}

\item{\code{data}}{any R object, the input data, for downstream analysis}
}}

\references{
\url{https://tingtingzhan-maxt.netlify.app/appendix/statistic.html#sec-free_d_math}
\url{https://tingtingzhan-maxt.netlify.app/S4/free.html#sec-free_d}
}
\keyword{internal}
