% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_internal.R
\name{free_internal}
\alias{free_internal}
\alias{.free_t}
\alias{.free_d}
\title{Create \linkS4class{free_t} Object}
\usage{
.free_t(x1, x0, null.value = 1, data, s4 = TRUE, ...)

.free_d(x1, x0, null.value = 0, data, s4 = TRUE, ...)
}
\arguments{
\item{x1}{\link[base]{numeric} \link[base]{matrix},
treatment responses \eqn{x_1}}

\item{x0}{\link[base]{numeric} \link[base]{matrix},
control responses \eqn{x_0}}

\item{null.value}{\link[base]{numeric} scalar, the constant \eqn{c},
default value is 1}

\item{data}{(optional) an R object,
the first argument of function \link{free_d} or \link{free_t},
for downstream analysis}

\item{s4}{\link[base]{logical} scalar, whether to return an \code{S4} object,
default value is \code{TRUE}. Set as \code{FALSE} for internal use to save compute time.}

\item{...}{additional parameters, currently of no use}
}
\description{
Low-level utility function to create \linkS4class{free_t} and \linkS4class{free_d} objects.
}
\references{
\url{https://tingtingzhan-maxt.netlify.app/S4/free_t.html#sec-free_t_create}
}
