% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_t_diff.R
\name{-,free_t,free_t-method}
\alias{-,free_t,free_t-method}
\title{Difference of \linkS4class{free_t} Objects}
\usage{
\S4method{-}{free_t,free_t}(e1, e2)
}
\arguments{
\item{e1, e2}{two \linkS4class{free_t} objects that have a common slot \verb{@null.value}.}
}
\description{
Difference of two \linkS4class{free_t} objects that have a common slot \verb{@null.value}.
}
\references{
\url{https://tingtingzhan-maxt.netlify.app/S4/free.html#sec-free_t_diff}
}
\keyword{internal}
