% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_t.R
\docType{class}
\name{free_t}
\alias{free_t}
\alias{free_t-class}
\alias{free_t.ELISpot}
\title{Distribution \linkS4class{free_t}-Statistic}
\usage{
free_t(x, ...)

\method{free_t}{ELISpot}(x, pid1, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{additional parameters, currently not in use}

\item{pid1}{(optional) \link[base]{integer} \link[base]{vector},
permuted indices for treatment}
}
\description{
Distribution \linkS4class{free_t} statistic to compare a treatment against a control.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\link[base]{numeric} \link[base]{vector}, distribution-free \eqn{t}-statistic}

\item{\code{null.value}}{\link[base]{numeric} scalar \eqn{c}}

\item{\code{delta}}{\link[base]{numeric} \link[base]{vector}}

\item{\code{stderr}}{\link[base]{numeric} \link[base]{vector}}

\item{\code{df}}{\link[base]{numeric} \link[base]{vector}}

\item{\code{data}}{any R object, the input data, for downstream analysis}
}}

\references{
\url{https://tingtingzhan-maxt.netlify.app/appendix/statistic.html#sec-free_t_math}
\url{https://tingtingzhan-maxt.netlify.app/s4/free_t#sec-free_t}
}
\keyword{internal}
