% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_t_diff.R
\docType{class}
\name{free_t_diff-class}
\alias{free_t_diff-class}
\title{Distribution Free \eqn{t}-Statistic, Difference of Difference}
\description{
The \code{S4} class \linkS4class{free_t_diff}.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\link[base]{numeric} \link[base]{vector}}

\item{\code{e1,e2}}{\linkS4class{free_t} objects}

\item{\code{design}}{\link[base]{data.frame}, study design showing the difference-of-difference}
}}

\references{
\url{https://tingtingzhan-maxt.netlify.app/S4/free_t.html#sec-free_t_diff}
}
\keyword{internal}
