% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_t.R
\docType{class}
\name{free_t_diff}
\alias{free_t_diff}
\alias{free_t_diff-class}
\title{Distribution Free Test Statistic, Difference of Difference \linkS4class{free_t_diff}}
\description{
Distribution Free Test Statistic, Difference of Difference \linkS4class{free_t_diff}
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\link[base]{numeric} \link[base]{vector}}

\item{\code{e1,e2}}{\linkS4class{free_t}}
}}

