% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_pvalue_sym.R
\name{label_pvalue_sym}
\alias{label_pvalue_sym}
\title{Label \eqn{p}-values with Significance Symbol}
\usage{
label_pvalue_sym(star = c("⭑", "⭒"), ...)
}
\arguments{
\item{star}{\link[base]{character} \link[base]{vector}
of \link[base]{length}-\code{2L} to denote the significance levels.
For example, to use \link[stats]{printCoefmat}-style of significance notation,
users should let \code{star = c('*', '.')}}

\item{...}{parameters of the function \link[scales]{label_pvalue}}
}
\value{
The function \code{\link[=label_pvalue_sym]{label_pvalue_sym()}} returns a \link[base]{function}.
}
\description{
Label \eqn{p}-values with significance symbol.
}
\details{
Pipeline:

Step 1: Apply function \link[scales]{label_pvalue}.

Step 2: Drop leading zeros.

Step 3: Apply function \link[stats]{symnum} (also see function \link[stats]{printCoefmat}).
}
\note{
The function \link[scales]{label_pvalue} is much prettier and more flexible than function \link[base]{format.pval}.
}
\examples{
p = c(a = pi^-100, b = .02, c = .05, d = .1, e = .9999, f = NA_real_)
p |> label_pvalue_sym()() # star
p |> label_pvalue_sym(star = c('\u2605', '\u2606'))() # small star
p |> label_pvalue_sym(star = c('*', '.'))() # ?stats::printCoefmat
p |> label_pvalue_sym(add_p = TRUE)()

# below: exception handling
double() |> scales::label_pvalue()() # do not like!
# ?scales::pvalue is bad; ?scales::number is fine!
double() |> label_pvalue_sym()() # nice!
}
\keyword{internal}
