% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.elispot.R
\name{log_elispot}
\alias{log_elispot}
\alias{log.elispot}
\alias{log1p.elispot}
\alias{log10.elispot}
\alias{log2.elispot}
\title{Several S3 method dispatches for \code{elispot}}
\usage{
\method{log}{elispot}(x, base = exp(1))

\method{log1p}{elispot}(x)

\method{log10}{elispot}(x)

\method{log2}{elispot}(x)
}
\arguments{
\item{x}{an \code{elispot} object}

\item{base}{\link[base]{numeric} scalar, see \link[base]{log}}
}
\value{
Function \code{\link[=log.elispot]{log.elispot()}} returns an \code{elispot}.
}
\description{
\code{elispot} dispatches for S3 generics \link[base]{log}.
}
\note{
The \code{S3} generic function \link[base]{log1p} does not have \code{base} parameter.
}
\keyword{internal}
