% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELISpot.R
\name{Math.ELISpot}
\alias{Math.ELISpot}
\title{\code{Math} \link[base]{groupGeneric} of \linkS4class{ELISpot}}
\usage{
\method{Math}{ELISpot}(x, ...)
}
\arguments{
\item{x}{an \linkS4class{ELISpot} object}

\item{...}{additional parameters}
}
\value{
Function \code{\link[=Math.ELISpot]{Math.ELISpot()}} returns an \linkS4class{ELISpot} object.
}
\description{
\code{Math} \link[base]{groupGeneric} of \linkS4class{ELISpot}
}
\keyword{internal}
