% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT_S4.R
\docType{class}
\name{maxT-class}
\alias{maxT-class}
\title{Westfall & Young's \linkS4class{maxT} Algorithm}
\description{
Westfall & Young's \linkS4class{maxT} algorithm, as described in Box 2, page 82 of \doi{10.1214/ss/1056397487}.
}
\section{Slots}{

\describe{
\item{\code{t.}}{\link[base]{length}-\eqn{m} \link[base]{double} \link[base]{vector},
test statistics \eqn{t_1,\cdots,t_m}
for each hypothesis \eqn{H_j}, \eqn{j = 1,\cdots,m},
in the original data}

\item{\code{T.}}{\link[base]{double} \link[base]{matrix} of \link[base]{dim}ension \eqn{(m,B)},
test statistics \eqn{t_{j,b}}
for permutation \eqn{b=1,\cdots,B}}

\item{\code{U}}{\link[base]{double} \link[base]{matrix} of \link[base]{dim}ension \eqn{(m,B)},
the successive maxima \eqn{u_{j,b}}}

\item{\code{p.value}}{\link[base]{length}-\eqn{m} \link[base]{double} \link[base]{vector},
\eqn{\tilde{p}_{r_j}} restored in the order of original hypotheses \eqn{1,\cdots,m}}

\item{\code{alternative}}{\link[base]{character} scalar, either
\code{'two.sided'} (default value), \code{'greater'} or \code{'less'}}

\item{\code{design}}{(optional) \link[base]{data.frame},
the study design}

\item{\code{label}}{(optional) \link[base]{character} scalar,
the study name}
}}

\note{
The algorithm is implemented in an unexported
\link[methods]{initialize} method, which could be revealed by
\code{getMethod(initialize, signature = 'maxT')} to curious eyes.
}
\references{
Peter H. Westfall, S. Stanley Young (1993). \emph{Resampling-Based Multiple Testing: Examples and Methods for p-Value Adjustment}. \url{https://www.wiley.com/en-us/Resampling-Based+Multiple+Testing\%3A+Examples+and+Methods+for+p-Value+Adjustment-p-9780471557616}

S. Dudoit, J. P. Shaffer, J. C. Boldrick (2003). \emph{Multiple Hypothesis Testing in Microarray Experiments},
\doi{10.1214/ss/1056397487}

\url{https://tingtingzhan-maxt.netlify.app/appendix/westfall_yang.html}
}
