% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT_.R
\name{maxT_}
\alias{maxT_}
\title{Westfall & Young's max\eqn{T} Algorithm}
\usage{
maxT_(t., T., two.sided = TRUE)
}
\arguments{
\item{t.}{\link[base]{numeric} \link[base]{vector} of
\eqn{t}-statistic \eqn{(t_1,\cdots,t_m)^t}
for each hypothesis \eqn{H_j, j = 1,\cdots,m}}

\item{T.}{\link[base]{numeric} \link[base]{matrix} of dimension \eqn{(m,B)},
permuted \eqn{t}-statistic \eqn{(t_{1,b},\cdots,t_{m,b})^t}
for permutation \eqn{b=1,\cdots,B}}

\item{two.sided}{\link[base]{logical} scalar,
whether to take \link[base]{abs}olute of \code{t.} and \code{T.}
before other calculations.
Default \code{TRUE} as in \doi{10.1214/ss/1056397487},
while Moodie's algorithm (\url{https://rundfr.fredhutch.org}) does not have this step.}
}
\description{
Westfall & Young's max\eqn{T} algorithm, as described in Box 2, page 82 of \doi{10.1214/ss/1056397487}
}
\references{
Peter H. Westfall, S. Stanley Young (1993). \emph{Resampling-Based Multiple Testing: Examples and Methods for p-Value Adjustment}. \url{https://www.wiley.com/en-us/Resampling-Based+Multiple+Testing\%3A+Examples+and+Methods+for+p-Value+Adjustment-p-9780471557616}

S. Dudoit, J. P. Shaffer, J. C. Boldrick (2003). \emph{Multiple Hypothesis Testing in Microarray Experiments},
\doi{10.1214/ss/1056397487}
}
