% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT_moodie.R
\name{maxT_moodie}
\alias{maxT_moodie}
\title{maxT_moodie}
\usage{
maxT_moodie(data, null.value, R = 1000L, two.sided = TRUE, ...)
}
\arguments{
\item{data}{an \linkS4class{ELISpot} object}

\item{null.value}{\link[base]{numeric} scalar \eqn{\mu_0}, as in \eqn{H_0: \log\mu_1 - \log\mu_0 = \mu_0}}

\item{R}{\link[base]{integer} scalar \eqn{R}, number of bootstrap copies if \code{bootstrap = TRUE}}

\item{two.sided}{\link[base]{logical} scalar, see \linkS4class{maxT}}

\item{...}{potential parameters}
}
\description{
Resampling-based \eqn{p}-value adjustment
}
\details{
..
}
\references{
Original algorithm by Zoe Moodie on March 16, 2010 \url{https://rundfr.fredhutch.org},
and
her 2006 papers \doi{10.1016/j.jim.2006.07.015},
and an empirical study \doi{10.1007/s00262-010-0875-4}.
}
\keyword{internal}
