% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT_santos.R
\name{maxT_santos}
\alias{maxT_santos}
\title{Modified Distribution Free Resampling}
\usage{
maxT_santos(data, ...)
}
\arguments{
\item{data}{an \code{ELISpot}}

\item{...}{additional parameters, such as \code{null.value} in function \link{santosT},
and \code{two.sided} for \linkS4class{maxT}}
}
\description{
..
}
\examples{
# paper page 5, 'The ELISPOT spot results were analyzed for statistical significance at 
#  .. a given time point by comparing the number of spots for each of the tested peptides 
#  .. (triplicates) with the background controls (no peptide added, six replicates) 
#  .. at that time point.'
# paper page 8, layout of Figure 2B 
# these are reasons why we use `f = ~ Hr + antigen`
st1 = split(santos1, f = ~ Hr + antigen)
st2 = split(santos2, f = ~ Hr + antigen)
maxT_santos(st1[[1L]], null.value = 1)
maxT_santos(st1[[1L]], null.value = 2)
lapply(st1, FUN = maxT_santos, null.value = 1)
lapply(st1, FUN = maxT_santos, null.value = 2)
lapply(st2, FUN = maxT_santos, null.value = 1)
lapply(st2, FUN = maxT_santos, null.value = 2)

}
\references{
Radleigh Santos's 2015 paper \doi{10.3390/cells4010001}
}
\keyword{internal}
