% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.maxT.R
\name{md_.maxT}
\alias{md_.maxT}
\title{md_.maxT}
\usage{
\method{md_}{maxT}(x, xnm, ...)
}
\arguments{
\item{x}{\linkS4class{maxT} object}

\item{xnm}{..}

\item{...}{..}
}
\description{
..
}
\examples{
library(rmd.tzh)
ds = split(santos1, f = ~ Hr + antigen)
list(
 '`maxT`' = maxT_santos_test(data1 = ds$`18.CEF`, data0 = ds$`0.CEF`)
) |> render_(file = 'maxT')

}
\keyword{internal}
