% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mDFR.R
\name{mDFR}
\alias{mDFR}
\title{Modified Distribution Free Resampling}
\usage{
mDFR(data, ...)
}
\arguments{
\item{data}{an \code{elispot}}

\item{...}{potential parameters}
}
\description{
..
}
\examples{
mDFR(santos1, null.value = 1)
mDFR(santos1, null.value = 2)

}
\references{
Radleigh Santos's 2015 paper \doi{10.3390/cells4010001}
}
