% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mDFR.R
\name{mDFR.test}
\alias{mDFR.test}
\title{Modified Distribution Free Resampling, at Two Timepoints}
\usage{
mDFR.test(data1, data0, null.value = 1, ...)
}
\arguments{
\item{data1, data0}{\code{elispot}s at two time points \eqn{t_1} and \eqn{t_0}, respectively}

\item{null.value}{..}

\item{...}{additional parameters, currently not in use}
}
\description{
..
}
