% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELISpot.R
\name{moodie2ELISpot}
\alias{moodie2ELISpot}
\title{Create \code{elispot} Object, Moodie's fashion}
\usage{
moodie2ELISpot(formula, data, control, ...)
}
\arguments{
\item{formula}{\link[stats]{formula}}

\item{data}{\link[base]{data.frame} with columns in the order of
patient_id, day, antigen, rep1, rep2, etc.}

\item{control}{\link[base]{character} scalar, name of negative control wells.
Currently do not allow \code{missing} (for no negative control wells)}

\item{...}{potential parameters, currently not in use}
}
\value{
Functions \code{\link[=moodie2ELISpot]{moodie2ELISpot()}} and \code{\link[=santos2ELISpot]{santos2ELISpot()}} both return an \code{elispot}.
}
\description{
..
}
\examples{
# see ?maxT_moodie
}
\references{
\url{https://rundfr.fredhutch.org}
}
\keyword{internal}
