% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{moodie_data}
\alias{moodie_data}
\alias{moodie}
\title{Moodie's data}
\format{
An object of class \code{data.frame} with 18 rows and 9 columns.
}
\usage{
moodie
}
\description{
Moodie's data
}
\keyword{internal}
