% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_elispot.R
\name{perm_elispot}
\alias{perm_elispot}
\title{Permutation Indices of Treatment in \code{elispot}}
\usage{
perm_elispot(data)
}
\arguments{
\item{data}{a \code{elispot} object}
}
\value{
Function \link{perm_elispot} returns a \link[base]{list} of \link[base]{integer} \link[base]{vector}s.
}
\description{
Permuted indices of treatment in a \code{elispot} object
}
\note{
We follow the terminology used in Westfall & Young's \linkS4class{maxT} algorithm
(Box 2, page 82 of \doi{10.1214/ss/1056397487}), which is \emph{permutation}.

In \strong{R} convention, this is indeed a \emph{combination} \link[utils]{combn}.
}
\keyword{internal}
