% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permID.R
\name{permID}
\alias{permID}
\title{Permuted Indices of Treatment}
\usage{
permID(x)
}
\arguments{
\item{x}{see \strong{Usage}}
}
\value{
The \code{S3} generic function \code{\link[=permID]{permID()}} returns a \link[base]{list} of \link[base]{integer} \link[base]{vector}s.
}
\description{
Permuted indices of treatment.
}
\note{
Westfall & Young's \linkS4class{maxT} algorithm
(Box 2, page 82 of \doi{10.1214/ss/1056397487}) used the term \emph{permutation}.

However, in \strong{R} convention, this is a \emph{combination} (via the function \link[utils]{combn}).
}
\keyword{internal}
