% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT.R
\name{rbind.maxT}
\alias{rbind.maxT}
\title{\code{\link[=rbind.maxT]{rbind.maxT()}}}
\usage{
\method{rbind}{maxT}(...)
}
\arguments{
\item{...}{..}
}
\description{
\code{\link[=rbind.maxT]{rbind.maxT()}}
}
\keyword{internal}
