% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{reactable_maxT}
\alias{reactable_maxT}
\title{Create \link[reactable]{reactable} from \linkS4class{maxT} Object}
\usage{
reactable_maxT(x, ...)
}
\arguments{
\item{x}{a \linkS4class{maxT} object}

\item{...}{additional parameters of function \link[reactable]{reactable}}
}
\value{
Function \code{\link[=reactable_maxT]{reactable_maxT()}} returns a \link[reactable]{reactable} object.
}
\description{
To create a \link[reactable]{reactable} from a \linkS4class{maxT} object.
}
