% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELISpot.R
\name{santos2ELISpot}
\alias{santos2ELISpot}
\title{Create \code{ELISpot} Object, Santos' fashion}
\usage{
santos2ELISpot(formula, data, ptn0 = "^y0_", ptn1 = "^y1_", ...)
}
\arguments{
\item{ptn0}{\link[base]{regex}, pattern of control}

\item{ptn1}{\link[base]{regex}, pattern of treatment}
}
\description{
..
}
\examples{
# these are `santos1` and `santos2` in this package
santos2ELISpot(formula = ~ antigen | Hr + Subj, data = santos1_raw)
santos2ELISpot(formula = ~ antigen | Hr + Subj, data = santos2_raw)

}
\keyword{internal}
