% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{santos_data}
\alias{santos_data}
\alias{santos1}
\alias{santos2}
\alias{santos1_raw}
\alias{santos2_raw}
\title{Santos' data}
\format{
An object of class \code{elispot} (inherits from \code{data.frame}) with 36 rows and 5 columns.

An object of class \code{elispot} (inherits from \code{data.frame}) with 36 rows and 5 columns.

An object of class \code{data.frame} with 36 rows and 12 columns.

An object of class \code{data.frame} with 36 rows and 12 columns.
}
\usage{
santos1

santos2

santos1_raw

santos2_raw
}
\description{
Santos' data
}
\keyword{datasets}
