% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/santosT.R
\name{-,santosT,santosT-method}
\alias{-,santosT,santosT-method}
\title{Distribution Free Test Statistic, Difference of Difference}
\usage{
\S4method{-}{santosT,santosT}(e1, e2)
}
\arguments{
\item{e1, e2}{\linkS4class{santosT}}
}
\description{
Distribution Free Test Statistic, Difference of Difference
}
\references{
The distribution free test statistic \eqn{T} is not exactly the same as
Equation (6) and (7) of Santos' 2015 cell paper \doi{10.3390/cells4010001}.
}
\keyword{internal}
