% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stat.R
\docType{class}
\name{santosT}
\alias{santosT}
\alias{santosT-class}
\alias{santosT.ELISpot}
\alias{santosT.matrix}
\title{Distribution Free Test Statistic}
\usage{
santosT(x, ...)

\method{santosT}{ELISpot}(x, ...)

\method{santosT}{matrix}(x, x0, data, null.value = 1, s4 = TRUE, ...)
}
\arguments{
\item{x, x0}{\link[base]{numeric} \link[base]{matrix}-es,
treatment \eqn{x_1} and control responses \eqn{x_0}, respectively}

\item{...}{additional parameters, currently not in use}

\item{null.value}{\link[base]{numeric} scalar \eqn{c},
as in \eqn{H_0: \mu_1 - c\mu_0 = 0}}

\item{s4}{\link[base]{logical} scalar}
}
\description{
Distribution free test statistic to compare a treatment against a control.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{..}

\item{\code{delta}}{..}

\item{\code{stderr}}{..}

\item{\code{df}}{..}

\item{\code{data}}{..}
}}

\references{
The distribution free test statistic \eqn{T} is not exactly equation (1) and (2) of Santos' 2015 cell paper \doi{10.3390/cells4010001}.
}
\keyword{internal}
