% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stat.R
\name{santosT2}
\alias{santosT2}
\title{Distribution Free Test Statistic, Difference of Difference}
\usage{
santosT2(u, v)
}
\arguments{
\item{u, v}{two objects, both returned from function \link{santosT}}
}
\value{
Function \code{\link[=santosT2]{santosT2()}} returns a \link[base]{numeric} \link[base]{vector} \eqn{T}.
}
\description{
Distribution Free Test Statistic, Difference of Difference
}
\references{
The distribution free test statistic \eqn{T} is not exactly the same as
Equation (6) and (7) of Santos' 2015 cell paper \doi{10.3390/cells4010001}.
}
