% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/santosT.R
\docType{class}
\name{santosT_diff}
\alias{santosT_diff}
\alias{santosT_diff-class}
\title{Distribution Free Test Statistic, Difference of Difference \linkS4class{santosT_diff}}
\description{
Distribution Free Test Statistic, Difference of Difference \linkS4class{santosT_diff}
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\link[base]{numeric} \link[base]{vector}}

\item{\code{e1,e2}}{\linkS4class{santosT}}
}}

