% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT_santos_test.R
\name{santosTm}
\alias{santosTm}
\title{Santos' \eqn{T}-statistic, between time point}
\usage{
santosTm(
  data1,
  data0,
  null.value = 1,
  y11 = data1$y1,
  y01 = data0$y1,
  y10 = data1$y0,
  y00 = data0$y0,
  ...
)
}
\arguments{
\item{data1, data0}{(optional) two \code{elispot} objects,
at two time points \eqn{t_1} and \eqn{t_0}, respectively}

\item{null.value}{\link[base]{numeric} scalar \eqn{c},
as in \eqn{H_0: (\bar{x}_{1,t_1} - c\cdot\bar{x}_{0,t_1}) - (\bar{x}_{1,t_0} - c\cdot\bar{x}_{0,t_0}) = 0}}

\item{y11}{\link[base]{numeric} \link[base]{matrix}, time \eqn{t_1} treatment response}

\item{y01}{\link[base]{numeric} \link[base]{matrix}, time \eqn{t_0} treatment response}

\item{y10}{\link[base]{numeric} \link[base]{matrix}, time \eqn{t_1} control response}

\item{y00}{\link[base]{numeric} \link[base]{matrix}, time \eqn{t_0} control response}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{santosTm} returns a \link[base]{numeric} \link[base]{vector}.
}
\description{
Equation (6) and (7) of Santos' 2015 cell paper \doi{10.3390/cells4010001}.
}
