% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT.R
\name{show,maxT-method}
\alias{show,maxT-method}
\title{\link[methods]{show} \linkS4class{maxT} object}
\usage{
\S4method{show}{maxT}(object)
}
\arguments{
\item{object}{a \linkS4class{maxT} object}
}
\value{
The \link[methods]{show} method of \linkS4class{maxT} object returns a \link[reactable]{reactable} object.
}
\description{
\link[methods]{show} \linkS4class{maxT} object
}
