% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{data}
\name{snook_data}
\alias{snook_data}
\alias{snook}
\title{Snook's data}
\format{
An object of class \code{data.frame} with 552 rows and 8 columns.
}
\usage{
snook
}
\description{
Snook's data
}
\keyword{internal}
