% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.elispot.R
\name{split.ELISpot}
\alias{split.ELISpot}
\title{\code{\link[=split.ELISpot]{split.ELISpot()}}}
\usage{
\method{split}{ELISpot}(x, f, drop = TRUE, ...)
}
\arguments{
\item{x}{an \linkS4class{ELISpot}}

\item{f}{..}

\item{drop}{\link[base]{logical} scalar, must be \code{FALSE}}
}
\description{
\code{\link[=split.ELISpot]{split.ELISpot()}}
}
\keyword{internal}
