% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT_S4.R
\name{[.maxT}
\alias{[.maxT}
\title{Subset of Hypothesis in Westfall & Young's \linkS4class{maxT} Algorithm}
\usage{
\method{[}{maxT}(x, i)
}
\arguments{
\item{x}{a \linkS4class{maxT} object}

\item{i}{\link[base]{logical} or \link[base]{integer} \link[base]{vector},
indices of a subset of hypothesis \eqn{\{i_1,\cdots,i_n\}\subset\{1,\cdots,m\}}}
}
\value{
Function \verb{[.maxT} returns a \linkS4class{maxT} object.
}
\description{
Subset of Hypothesis in Westfall & Young's \linkS4class{maxT} Algorithm
}
\details{
Function \verb{[.maxT} performs Westfall & Young's \linkS4class{maxT} algorithm
on a subset of test statistics \eqn{\{t_{i_1},\cdots t_{i_n}\}\subset\{t_1,\cdots t_m\}}
and their corresponding test statistics
\eqn{\{t_{i_1,b},\cdots,t_{i_n,b}\}\subset\{t_{1,b},\cdots,t_{m,b}\}}
in each permuted copy \eqn{b}, \eqn{b=1,\cdots,B}.

In other words, we take the subset of hypothesis \eqn{j=1,\cdots,m},
\strong{not} of permutations copies \eqn{b=1,\cdots,B}.
}
\keyword{internal}
