% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_elispot.R
\name{trim_ELISpot}
\alias{trim_ELISpot}
\title{Trim \code{'ELISpot'} Object}
\usage{
trim_ELISpot(x)
}
\arguments{
\item{x}{an \code{ELISpot} object}
}
\value{
Function \code{\link[=trim_ELISpot]{trim_ELISpot()}} returns an \code{ELISpot} object.
}
\description{
Trim \code{'ELISpot'} Object
}
\details{
Both treatment \verb{@x1} and control \verb{@x0} must contain more than 1 measurement
}
\keyword{internal}
