% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT.R
\name{as.data.frame.maxT}
\alias{as.data.frame.maxT}
\title{Convert \linkS4class{maxT} to a \link[base]{data.frame}}
\usage{
\method{as.data.frame}{maxT}(x, ...)
}
\arguments{
\item{x}{a \linkS4class{maxT} object}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{as.data.frame.maxT} returns a \link[base]{data.frame}.
}
\description{
Convert \linkS4class{maxT} to a \link[base]{data.frame}
}
