% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELISpot.R
\name{as.ELISpot}
\alias{as.ELISpot}
\title{Create \code{ELISpot} Object, Moodie's fashion}
\usage{
as.ELISpot(formula, data, pattern0, pattern1, ...)
}
\arguments{
\item{formula}{\link[stats]{formula}}

\item{data}{\link[base]{data.frame} with columns in the order of
patient_id, day, antigen, rep1, rep2, etc.}

\item{pattern0}{\link[base]{character} scalar, name of negative control wells.
Currently do not allow \code{missing} (for no negative control wells)}

\item{pattern1}{\link[base]{regex}}

\item{...}{potential parameters, currently not in use}
}
\value{
Functions \code{\link[=santos2ELISpot]{santos2ELISpot()}} both return an \code{ELISpot}.
}
\description{
..
}
\keyword{internal}
