% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT.R
\name{autoplot.maxT}
\alias{autoplot.maxT}
\title{autoplot.maxT}
\usage{
\method{autoplot}{maxT}(object, vertical = TRUE, ...)
}
\arguments{
\item{object}{\linkS4class{maxT}}

\item{vertical}{\link[base]{logical} scalar}

\item{...}{..}
}
\description{
autoplot.maxT
}
