% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT.R
\name{autoplot.maxT}
\alias{autoplot.maxT}
\title{Visualize Westfall & Young's \linkS4class{maxT} Algorithm}
\usage{
\method{autoplot}{maxT}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{a \linkS4class{maxT} object}

\item{conf.level}{\link[base]{double} scalar,
confidence level, or \eqn{1-\alpha}. Default .95 (or \eqn{\alpha=.05})}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=autoplot.maxT]{autoplot.maxT()}} returns a \link[ggplot2]{ggplot} object.
}
\description{
To visualize Westfall & Young's \linkS4class{maxT} algorithm using package \CRANpkg{ggplot2}.
}
\details{
Function \code{\link[=autoplot.maxT]{autoplot.maxT()}} plots
\itemize{
\item the successive maxima
\eqn{u_{jb}}, \eqn{j=1,\cdots,m}, \eqn{b=1,\cdots,B}, and
\item the decreasing-ordered statistics \eqn{|t_{r_1}|\geq|t_{r_2}|\geq\cdots\geq|t_{r_m}|} for two-sided test,
or \eqn{t_{r_1}\geq t_{r_2}\geq\cdots\geq t_{r_m}} for one-sided test
}
Printed on opposing axis are
\itemize{
\item values of the decreasing-ordered statistics
\item permutation adjusted \eqn{p}-values \eqn{\tilde{p}_{r_j}}, as well as under monotonicity constraints \eqn{\tilde{p}^*_{r_j}}
}
Tests with \eqn{\tilde{p}^*_{r_j}\leq\alpha} is considered significant
and colored pink (hex color \verb{#F8766D}), otherwise non-significant and colored blue (hex color \verb{#00BFC4})

See full details of these notations in \linkS4class{maxT}.
}
\keyword{internal}
