% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELISpot.R
\docType{class}
\name{ELISpot}
\alias{ELISpot}
\alias{ELISpot-class}
\title{S4 class \linkS4class{ELISpot}}
\description{
S4 class \linkS4class{ELISpot}
}
\section{Slots}{

\describe{
\item{\code{design}}{\link[base]{data.frame}, experimental design}

\item{\code{x1}}{\link[base]{matrix}}

\item{\code{x0}}{\link[base]{matrix}}
}}

\keyword{internal}
