% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELISpot.R
\name{ELISpot}
\alias{ELISpot}
\alias{moodie2ELISpot}
\alias{santos2ELISpot}
\title{Create \code{elispot} Object}
\usage{
moodie2ELISpot(formula, data, control, ...)

santos2ELISpot(formula, data, ptn0 = "^y0_", ptn1 = "^y1_", ...)
}
\arguments{
\item{formula}{\link[stats]{formula}}

\item{data}{\link[base]{data.frame} with columns in the order of
patient_id, day, antigen, rep1, rep2, etc.}

\item{control}{\link[base]{character} scalar, name of negative control wells.
Currently do not allow \code{missing} (for no negative control wells)}

\item{...}{potential parameters, currently not in use}

\item{ptn0}{\link[base]{regex}, pattern of control}

\item{ptn1}{\link[base]{regex}, pattern of treatment}
}
\value{
Functions \link{moodie2ELISpot} and \link{santos2ELISpot} both return an \code{elispot}.
}
\description{
..
}
\examples{
# ELISpot_Moodie(moodie, nameCtrl = 'negctl') # deprecated
moodie2ELISpot(`^a[1-9]$` ~ antigen | day + id, data = moodie, control = 'negctl')
# these are `santos1` and `santos2` in this package
santos2ELISpot(formula = ~ antigen | Hr + Subj, data = santos1_raw)
santos2ELISpot(formula = ~ antigen | Hr + Subj, data = santos2_raw)

}
\references{
\url{https://rundfr.fredhutch.org}
}
