% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELISpot.R
\name{elispot_S3}
\alias{elispot_S3}
\alias{log.elispot}
\title{Several S3 method dispatches for \code{elispot}}
\usage{
\method{log}{elispot}(x, base = exp(1))
}
\arguments{
\item{x}{an \code{elispot} object}

\item{base}{\link[base]{numeric} scalar, see \link[base]{log}}
}
\value{
Function \link{log.elispot} returns an \code{elispot}.
}
\description{
\code{elispot} dispatches for S3 generics \link[base]{log}, \link[base]{split}.
}
\note{
Unfortunately \link[base]{log10} and \link[base]{log1p} are not an S3 generics.

\link[base]{log1p} does not have \code{base} parameter.
}
