% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_d.R
\docType{class}
\name{free_d}
\alias{free_d}
\alias{free_d-class}
\alias{free_d.ELISpot}
\alias{free_d.matrix}
\title{Distribution \linkS4class{free_d}-Statistic}
\usage{
free_d(x, ...)

\method{free_d}{ELISpot}(x, id1, ...)

\method{free_d}{matrix}(x, x0, data, null.value = 0, s4 = TRUE, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{additional parameters, currently not in use}

\item{id1}{(optional) \link[base]{integer} \link[base]{vector},
permuted indices for treatment}

\item{x0}{\link[base]{numeric} \link[base]{matrix},
control responses \eqn{x_0}}

\item{null.value}{see \strong{Slots}}

\item{s4}{\link[base]{logical} scalar}
}
\description{
Distribution \linkS4class{free_d} statistic to compare a treatment against a control.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{..}

\item{\code{null.value}}{\link[base]{numeric} scalar \eqn{c},
as in \eqn{H_0: \mu_1 - c\mu_0 = 0}}

\item{\code{data}}{..}
}}

\keyword{internal}
