% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_t.R
\docType{class}
\name{free_t}
\alias{free_t}
\alias{free_t-class}
\alias{free_t.ELISpot}
\alias{free_t.matrix}
\title{Distribution Free Test Statistic \linkS4class{free_t}}
\usage{
free_t(x, ...)

\method{free_t}{ELISpot}(x, id1, ...)

\method{free_t}{matrix}(x, x0, data, null.value = 1, s4 = TRUE, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{additional parameters, currently not in use}

\item{id1}{(optional) \link[base]{integer} \link[base]{vector},
permuted indices for treatment}

\item{x0}{\link[base]{numeric} \link[base]{matrix},
control responses \eqn{x_0}}

\item{null.value}{see \strong{Slots}}

\item{s4}{\link[base]{logical} scalar}
}
\description{
Distribution free test statistic to compare a treatment against a control.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{..}

\item{\code{null.value}}{\link[base]{numeric} scalar \eqn{c},
as in \eqn{H_0: \mu_1 - c\mu_0 = 0}}

\item{\code{delta}}{..}

\item{\code{stderr}}{..}

\item{\code{df}}{..}

\item{\code{data}}{..}
}}

\references{
The distribution free test statistic \eqn{T} is not exactly equation (1) and (2) of Santos' 2015 cell paper \doi{10.3390/cells4010001}.
}
\keyword{internal}
