% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.elispot.R
\name{log_ELISpot}
\alias{log_ELISpot}
\alias{log.ELISpot}
\alias{log1p.ELISpot}
\alias{log10.ELISpot}
\alias{log2.ELISpot}
\title{Several S3 method dispatches for \code{ELISpot}}
\usage{
\method{log}{ELISpot}(x, base = exp(1))

\method{log1p}{ELISpot}(x)

\method{log10}{ELISpot}(x)

\method{log2}{ELISpot}(x)
}
\arguments{
\item{x}{an \code{ELISpot} object}

\item{base}{\link[base]{numeric} scalar, see \link[base]{log}}
}
\value{
Function \code{\link[=log.ELISpot]{log.ELISpot()}} returns an \code{ELISpot}.
}
\description{
\code{ELISpot} dispatches for S3 generics \link[base]{log}.
}
\note{
The \code{S3} generic function \link[base]{log1p} does not have \code{base} parameter.
}
\keyword{internal}
