% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT_S4.R
\docType{class}
\name{maxT-class}
\alias{maxT-class}
\title{Westfall & Young's \linkS4class{maxT} Algorithm}
\description{
Westfall & Young's \linkS4class{maxT} algorithm, as described in Box 2, page 82 of \doi{10.1214/ss/1056397487}.
}
\details{
In the original data, obtain the test statistics \eqn{t_1,\cdots,t_m}
for each hypothesis \eqn{H_j}, \eqn{j=1,\cdots,m},
as well as the \emph{decreasing order statistics} \eqn{\mathbf{r} = (r_1,\cdots,r_m)^t}, such that
\deqn{
\begin{cases}
|t_{r_1}|\geq|t_{r_2}|\geq\cdots\geq|t_{r_m}| & \text{for two-sided test}\\
t_{r_1}\geq t_{r_2}\geq\cdots\geq t_{r_m} & \text{for one-sided test}
\end{cases}
}

In each permuted copy \eqn{b}, \eqn{b=1,\cdots,B},
obtain the test statistics \eqn{t_{1,b},\cdots,t_{m,b}}
and their \strong{successive maxima} for two-sided test
\deqn{
\begin{cases}
u_{m,b} & = |t_{r_m,b}|\\
u_{j,b} & = \text{max}\left\{u_{j+1,b}, |t_{r_j,b}|\right\}, \quad \text{for}\ j = m−1,\cdots,1\\
\end{cases}
}
or for one-sided test
\deqn{
\begin{cases}
u_{m,b} & = t_{r_m,b}\\
u_{j,b} & = \text{max}\left\{u_{j+1,b}, t_{r_j,b}\right\}, \quad \text{for}\ j = m−1,\cdots,1\\
\end{cases}
}

The permutation adjusted \eqn{p}-values are
\deqn{
\tilde{p}_{r_j} = 
\begin{cases}
\dfrac{1}{B}\displaystyle\sum_{b=1}^B I\left(u_{j,b}\geq|t_{r_j}|\right) & \text{for two-sided test}\\
\dfrac{1}{B}\displaystyle\sum_{b=1}^B I\left(u_{j,b}\geq t_{r_j}\right) & \text{for one-sided test}
\end{cases}
}

To enforce the monotonicity constraints,
\deqn{
\begin{cases}
\tilde{p}^*_{r_1} & = \tilde{p}_{r_1}\\
\tilde{p}^*_{r_j} & = \text{max}\left\{\tilde{p}_{r_j}, \tilde{p}^*_{r_j-1}\right\}, \quad \text{for}\ j = 2,\cdots,m\\
\end{cases}
}
}
\section{Slots}{

\describe{
\item{\code{t.}}{\link[base]{double} \link[base]{vector},
test statistics \eqn{t_1,\cdots,t_m}
for each hypothesis \eqn{H_j}, \eqn{j = 1,\cdots,m},
in the original data}

\item{\code{T.}}{\link[base]{double} \link[base]{matrix} of dimension \eqn{(m,B)},
test statistics \eqn{t_{j,b}}, \eqn{j = 1,\cdots,m},
for permutation \eqn{b=1,\cdots,B}}

\item{\code{tr}}{\link[base]{double} \link[base]{vector}, ordered test statistics \eqn{t_{r_1}\geq t_{r_2}\geq\cdots\geq t_{r_m}} for one-sided test, or \eqn{|t_{r_1}|\geq|t_{r_2}|\geq\cdots\geq|t_{r_m}|} for two-sided test}

\item{\code{U}}{\link[base]{double} \link[base]{matrix} of dimension \eqn{(m,B)}, successive maxima \eqn{u_{j,b}}, \eqn{j=1,\cdots,m}, \eqn{b=1,\cdots,B}}

\item{\code{p_perm}}{\link[base]{double} \link[base]{vector}, permutation adjusted \eqn{p}-values \eqn{\tilde{p}_{r_j}}}

\item{\code{p_mono}}{\link[base]{double} \link[base]{vector}, permutation adjusted \eqn{p}-values under monotonicity constraints \eqn{\tilde{p}^*_{r_j}}}

\item{\code{p.}}{\link[base]{double} \link[base]{vector}, \eqn{\tilde{p}^*_{r_j}} restored in the order of original test statistics \eqn{t_1,\cdots,t_m}}

\item{\code{two.sided}}{two.sided \link[base]{logical} scalar,
whether to perform a two sided test.
Default \code{TRUE} as in \doi{10.1214/ss/1056397487},
while Moodie's (\url{https://rundfr.fredhutch.org}) use \code{FALSE}}

\item{\code{design}}{(optional) \link[base]{data.frame}, study design}

\item{\code{name}}{(optional) \link[base]{character} scalar, study name}
}}

\note{
The altorithm described in \strong{Details} is implemented in an unexported
\link[methods]{initialize} method, which could be revealed by
\code{getMethod(initialize, signature = 'maxT')} to curious eyes.
}
\references{
Peter H. Westfall, S. Stanley Young (1993). \emph{Resampling-Based Multiple Testing: Examples and Methods for p-Value Adjustment}. \url{https://www.wiley.com/en-us/Resampling-Based+Multiple+Testing\%3A+Examples+and+Methods+for+p-Value+Adjustment-p-9780471557616}

S. Dudoit, J. P. Shaffer, J. C. Boldrick (2003). \emph{Multiple Hypothesis Testing in Microarray Experiments},
\doi{10.1214/ss/1056397487}
}
