% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT.R
\name{maxT}
\alias{maxT}
\alias{maxT.free_d}
\alias{maxT.free_t}
\alias{maxT.free_t_diff}
\title{\linkS4class{maxT} of Distribution Free Statistics}
\usage{
maxT(x, ...)

\method{maxT}{free_d}(x, ...)

\method{maxT}{free_t}(x, ...)

\method{maxT}{free_t_diff}(x, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{additional parameters, currently not in use}
}
\description{
..
}
\references{
\url{https://tingtingzhan-maxt.netlify.app/appendix/westfall_yang.html}
}
\keyword{internal}
