% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT.R
\name{maxT}
\alias{maxT}
\alias{maxT.santosT}
\alias{maxT.santosT_diff}
\title{Modified Distribution Free Resampling}
\usage{
maxT(x, two.sided, ...)

\method{maxT}{santosT}(x, two.sided = TRUE, ...)

\method{maxT}{santosT_diff}(x, two.sided = TRUE, ...)
}
\arguments{
\item{x}{an \linkS4class{ELISpot}}

\item{two.sided}{\link[base]{logical} scalar, see \linkS4class{maxT}}

\item{...}{additional parameters, currently not in use}
}
\description{
..
}
\references{
Radleigh Santos's 2015 paper \doi{10.3390/cells4010001}
}
\keyword{internal}
