% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT.R
\name{maxT}
\alias{maxT}
\alias{maxT.santosT}
\alias{maxT.santosT_diff}
\title{Modified Distribution Free Resampling}
\usage{
maxT(x, ...)

\method{maxT}{santosT}(x, ...)

\method{maxT}{santosT_diff}(x, ...)
}
\arguments{
\item{x}{an \linkS4class{ELISpot}}

\item{...}{additional parameters, such as \code{null.value} in function \link{santosT},
and \code{two.sided} for \linkS4class{maxT}}
}
\description{
..
}
\references{
Radleigh Santos's 2015 paper \doi{10.3390/cells4010001}
}
\keyword{internal}
