% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT.R
\docType{class}
\name{maxT}
\alias{maxT}
\alias{maxT-class}
\title{Westfall & Young's max\eqn{T} Algorithm}
\usage{
maxT(t., T., two.sided = TRUE)
}
\arguments{
\item{t.}{\link[base]{numeric} \link[base]{vector},
test statistic \eqn{(t_1,\cdots,t_m)^t}
for each hypothesis \eqn{H_j, j = 1,\cdots,m},
in the original data}

\item{T.}{\link[base]{numeric} \link[base]{matrix} of dimension \eqn{(m,B)},
test statistic \eqn{(t_{1,b},\cdots,t_{m,b})^t}
for permutation \eqn{b=1,\cdots,B}}

\item{two.sided}{\link[base]{logical} scalar,
whether to take \link[base]{abs}olute of \code{t.} and \code{T.}
before other calculations.
Default \code{TRUE} as in \doi{10.1214/ss/1056397487},
while Moodie's algorithm (\url{https://rundfr.fredhutch.org}) does not have this step.}
}
\value{
Function \link{maxT} returns an S4 \linkS4class{maxT} object.  See slots in section \strong{Slots}.
}
\description{
Westfall & Young's max\eqn{T} algorithm, as described in Box 2, page 82 of \doi{10.1214/ss/1056397487}
}
\details{
In the original data, obtain the test statistics \eqn{t_1,\cdots,t_m}
for each hypothesis \eqn{H_j}, \eqn{j=1,\cdots,m},
as well as the \emph{decreasing order statistic} \eqn{\mathbf{r} = (r_1,\cdots,r_m)^t}, such that
\deqn{|t_{r_1}|\geq|t_{r_2}|\geq\cdots\geq|t_{r_m}|}

In each permuted copy \eqn{b}, \eqn{b=1,\cdots,B},
obtain the test statistics \eqn{t_{1,b},\cdots,t_{m,b}}
and their \strong{successive maxima} by the decreasing order statistic \eqn{\mathbf{r}} of the original data
\deqn{
\begin{cases}
u_{m,b} & = |t_{r_m,b}|\\
u_{j,b} & = \text{max}\left\{u_{j+1,b}, |t_{r_j,b}|\right\}, \quad \text{for}\ j = m−1,\cdots,1\\
\end{cases}
}

The permutation adjusted \eqn{p}-values are
\deqn{
\tilde{p}_{r_j} = B^{-1}\sum_{b=1}^B I\left(u_{j,b}\geq|t_{r_j}|\right)
}

To enforce the monotonicity constraints,
\deqn{
\begin{cases}
\tilde{p}^*_{r_1} & = \tilde{p}_{r_1}\\
\tilde{p}^*_{r_j} & = \text{max}\left\{\tilde{p}_{r_j}, \tilde{p}^*_{r_j-1}\right\}, \quad \text{for}\ j = 2,\cdots,m\\
\end{cases}
}
}
\section{Slots}{

\describe{
\item{\code{t.}}{\link[base]{double} \link[base]{vector}, original test statistics \eqn{t_1,\cdots,t_m}}

\item{\code{tr}}{\link[base]{double} \link[base]{vector}, ordered test statistics \eqn{t_{r_1}\geq t_{r_2}\geq\cdots\geq t_{r_m}} if \code{!two.sided}, or \eqn{|t_{r_1}|\geq|t_{r_2}|\geq\cdots\geq|t_{r_m}|} if \code{two.sided}}

\item{\code{U}}{\link[base]{double} \link[base]{matrix} of dimension \eqn{(m,B)}, successive maxima \eqn{u_{j,b}}, \eqn{j=1,\cdots,m}, \eqn{b=1,\cdots,B}}

\item{\code{p_perm}}{\link[base]{double} \link[base]{vector}, permutation adjusted \eqn{p}-values \eqn{\tilde{p}_{r_j}}}

\item{\code{p_mono}}{\link[base]{double} \link[base]{vector}, permutation adjusted \eqn{p}-values under monotonicity constraints \eqn{\tilde{p}^*_{r_j}}}

\item{\code{p.}}{\link[base]{double} \link[base]{vector}, permutation adjusted \eqn{p}-values under monotonicity constraints, restored in the order of original test statistics \eqn{t_1,\cdots,t_m}}

\item{\code{two.sided}}{\link[base]{logical} scalar}

\item{\code{design}}{(optional) \link[base]{data.frame}, study design}

\item{\code{name}}{(optional) \link[base]{character} scalar, study name}
}}

\references{
Peter H. Westfall, S. Stanley Young (1993). \emph{Resampling-Based Multiple Testing: Examples and Methods for p-Value Adjustment}. \url{https://www.wiley.com/en-us/Resampling-Based+Multiple+Testing\%3A+Examples+and+Methods+for+p-Value+Adjustment-p-9780471557616}

S. Dudoit, J. P. Shaffer, J. C. Boldrick (2003). \emph{Multiple Hypothesis Testing in Microarray Experiments},
\doi{10.1214/ss/1056397487}
}
