% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT_moodie.R
\name{maxT_moodie}
\alias{maxT_moodie}
\title{maxT_moodie}
\usage{
maxT_moodie(
  data,
  resample = c("combn", "boot"),
  null.value,
  R = 1000L,
  seed_,
  ...
)
}
\arguments{
\item{data}{an \code{elispot} object}

\item{resample}{\link[base]{character} scalar, \code{'combn'} (default) or \code{'boot'}}

\item{null.value}{\link[base]{numeric} scalar \eqn{\mu_0}, as in \eqn{H_0: \bar{x}_1 - \bar{x}_0 = \mu_0}}

\item{R}{\link[base]{integer} scalar \eqn{R}, number of bootstrap copies if \code{resample = 'boot'}}

\item{seed_}{(optional) \link[base]{integer} scalar, random seed for \code{'boot'} resampling method.
In Moodie's function \code{elsdfr2x()}, \code{seed_ = 9456845L} is used.
Howwever, Moodie uses a same seed for all analysis, which is a \emph{bad} practice!
The purpose of this parameter is solely to re-create Moodie's results.
The end user is advised to leave this parameter missing.}

\item{...}{potential parameters}
}
\description{
Resampling-based \eqn{p}-value adjustment
}
\details{
..
}
\references{
Original algorithm by Zoe Moodie on March 16, 2010, and
her 2006 papers \doi{10.1016/j.jim.2006.07.015},
and an empirical study \doi{10.1007/s00262-010-0875-4}.

\url{https://rundfr.fredhutch.org}
}
