% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT_santos.R
\name{maxT_santos}
\alias{maxT_santos}
\title{Modified Distribution Free Resampling}
\usage{
maxT_santos(data, null.value = 1, ...)
}
\arguments{
\item{data}{an \code{elispot}}

\item{null.value}{..}

\item{...}{potential parameters}
}
\description{
..
}
\examples{
# paper page 5, 'The ELISPOT spot results were analyzed for statistical significance at 
#  .. a given time point by comparing the number of spots for each of the tested peptides 
#  .. (triplicates) with the background controls (no peptide added, six replicates) 
#  .. at that time point.'
# paper page 8, layout of Figure 2B 
# these are reasons why we use `f = ~ Hr + antigen`
lapply(split(santos1, f = ~ Hr + antigen), FUN = maxT_santos, null.value = 1)
lapply(split(santos1, f = ~ Hr + antigen), FUN = maxT_santos, null.value = 2)
lapply(split(santos2, f = ~ Hr + antigen), FUN = maxT_santos, null.value = 1)
lapply(split(santos2, f = ~ Hr + antigen), FUN = maxT_santos, null.value = 2)

}
\references{
Radleigh Santos's 2015 paper \doi{10.3390/cells4010001}
}
