% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT_santos_test.R
\name{maxT_santos_test}
\alias{maxT_santos_test}
\title{Modified Distribution Free Resampling, at Two Timepoints}
\usage{
maxT_santos_test(data1, data0, ...)
}
\arguments{
\item{data1, data0}{two \code{elispot} objects,
at two time points \eqn{t_1} and \eqn{t_0}, respectively}

\item{...}{additional parameters, such as \code{null.value} in function \link{santosT} and
\code{two.sided} in function \link{maxT}}
}
\description{
..
}
\examples{
# to recreate Figure 2B
ds = split(santos1, f = ~ Hr + antigen)
maxT_santos_test(data1 = ds$`18.CEF`, data0 = ds$`0.CEF`)
maxT_santos_test(data1 = ds$`18.CEF`, data0 = ds$`0.CEF`, null.value = 10)
maxT_santos_test(data1 = ds$`18.CEF`, data0 = ds$`0.CEF`, two.sided = FALSE)

}
