% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT_santos_test.R
\name{maxT_santos_test}
\alias{maxT_santos_test}
\title{Modified Distribution Free Resampling, at Two Timepoints}
\usage{
maxT_santos_test(data1, data0, null.value = 1, ...)
}
\arguments{
\item{data1, data0}{\code{elispot}s at two time points \eqn{t_1} and \eqn{t_0}, respectively}

\item{null.value}{..}

\item{...}{additional parameters, currently not in use}
}
\description{
..
}
\examples{
# to recreate Figure 2B
ds = split(santos1, f = ~ Hr + antigen)
maxT_santos_test(data1 = ds$`18.CEF`, data0 = ds$`0.CEF`)
maxT_santos_test(data1 = ds$`18.CEF`, data0 = ds$`0.CEF`, two.sided = FALSE)

}
