% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{data}
\name{santos_data}
\alias{santos_data}
\alias{santos1}
\alias{santos2}
\title{Santos' data}
\format{
An object of class \code{data.frame} with 36 rows and 12 columns.

An object of class \code{data.frame} with 36 rows and 12 columns.
}
\usage{
santos1

santos2
}
\description{
Santos' data
}
\keyword{datasets}
\keyword{internal}
