% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stat.R
\name{santosT}
\alias{santosT}
\title{Distribution Free Test Statistic}
\usage{
santosT(x1, x0, null.value = 1, ...)
}
\arguments{
\item{x1, x0}{\link[base]{numeric} \link[base]{matrix}-es,
treatment \eqn{x_1} and control responses \eqn{x_0}, respectively}

\item{null.value}{\link[base]{numeric} scalar \eqn{c},
as in \eqn{H_0: \mu_1 - c\mu_0 = 0}}

\item{...}{additional parameters, currently not in use}
}
\description{
Distribution free test statistic to compare a treatment against a control.
}
\references{
The distribution free test statistic \eqn{T} is not exactly equation (1) and (2) of Santos' 2015 cell paper \doi{10.3390/cells4010001}.
}
\keyword{internal}
