% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mDFR.R
\name{santosT}
\alias{santosT}
\title{Santos' \eqn{T}-statistic, at given time point}
\usage{
santosT(y1, y0, null.value)
}
\arguments{
\item{y1, y0}{\link[base]{numeric} \link[base]{matrix}-es, treatment and control responses, respectively}

\item{null.value}{\link[base]{numeric} scalar \eqn{\mu_0}, as in \eqn{H_0: \bar{x}_1 - \mu_0\bar{x}_0 = 0}}
}
\value{
Function \link{santosT} returns a \link[base]{numeric} \link[base]{vector}.
}
\description{
Equation (1) and (2) of Santos' 2015 cell paper \doi{10.3390/cells4010001}.
}
