% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT_santos.R
\name{santosT}
\alias{santosT}
\title{Santos' Test-Statistic, at given time point}
\usage{
santosT(y1, y0, null.value = 1, ...)
}
\arguments{
\item{y1, y0}{\link[base]{numeric} \link[base]{matrix}-es,
treatment and control responses (from an \code{elispot} object), respectively}

\item{null.value}{\link[base]{numeric} scalar \eqn{c},
as in \eqn{H_0: \bar{x}_1 - c\cdot\bar{x}_0 = 0}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{santosT} returns a \link[base]{numeric} \link[base]{vector}.
}
\description{
Equation (1) and (2) of Santos' 2015 cell paper \doi{10.3390/cells4010001}.
}
