% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stat.R
\name{santosT2}
\alias{santosT2}
\title{Distribution Free Test Statistic, Difference of Difference}
\usage{
santosT2(u, v)
}
\arguments{
\item{u, v}{two objects, both returned from function \link{santosT}}
}
\value{
Function \code{\link[=santosT2]{santosT2()}} returns a \link[base]{numeric} \link[base]{vector} \eqn{T}.
}
\description{
Distribution Free Test Statistic, Difference of Difference
}
\details{
At time point \eqn{u} (see more at \link{santosT}), distribution free test statistic
\deqn{
T_u = \dfrac{\bar{X}_{1,u} - c_u\bar{X}_{0,u}}{s_{\bar\Delta,u}}
}
The standard deviation
\eqn{s^2_{\bar\Delta,u}=s^2_{1,u}/n_{1,u} + c_u^2s^2_{0,u}/n_{0,u}}
has degree of freedom based on Welch–Satterthwaite equation
\deqn{\text{df}_u = \dfrac{\left(\dfrac{s^2_{1,u}}{n_{1,u}}+\dfrac{c_u^2s^2_{0,u}}{n_{0,u}}\right)^2}{\dfrac{(s^2_{1,u}/n_{1,u})^2}{n_{1,u}-1}+\dfrac{(c_u^2s^2_{0,u}/n_{0,u})^2}{n_{0,u}-1}}}
Similarly, at time point \eqn{v}, we have distribution free test statistic \eqn{T_v},
standard deviation \eqn{s^2_{\bar\Delta,v}} and degree of freedom \eqn{\text{df}_v}.

To compare time point \eqn{u} vs. \eqn{v}, we have distribution free test statistic
\deqn{
T = \dfrac{\left(\bar{X}_{1,u} - c_u\bar{X}_{0,u}\right) - \left(\bar{X}_{1,v} - c_v\bar{X}_{0,v}\right)}{s_{\bar\Delta}}
}
where
\deqn{
s^2_{\bar\Delta} = \dfrac{\text{df}_u\cdot s^2_{\bar\Delta,u} + \text{df}_v\cdot s^2_{\bar\Delta,v}}{\text{df}_u + \text{df}_v}
}
}
\note{
In practice we often let \eqn{c=c_u=c_v} for more intuitive interpretation.
}
\references{
The distribution free test statistic \eqn{T} is not exactly the same as
Equation (6) and (7) of Santos' 2015 cell paper \doi{10.3390/cells4010001}.
}
