% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxT_santos_test.R
\name{santosTm}
\alias{santosTm}
\title{Santos' \eqn{T}-statistic, between time point}
\usage{
santosTm(tm1, tm0)
}
\arguments{
\item{tm1, tm0}{returns from function \link{santosT} at two time points \eqn{t_1} and \eqn{t_2}}
}
\value{
Function \link{santosTm} returns a \link[base]{numeric} \link[base]{vector}.
}
\description{
Equation (6) and (7) of Santos' 2015 cell paper \doi{10.3390/cells4010001}.
}
