% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELISpot.R
\name{split.ELISpot}
\alias{split.ELISpot}
\title{\link[base]{split} an \linkS4class{ELISpot} Object}
\usage{
\method{split}{ELISpot}(x, f, drop = TRUE, ...)
}
\arguments{
\item{x}{an \linkS4class{ELISpot}}

\item{f}{..}

\item{drop}{\link[base]{logical} scalar, must be \code{FALSE}}
}
\description{
\link[base]{split} an \linkS4class{ELISpot} Object
}
\keyword{internal}
