% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.elispot.R
\name{split.elispot}
\alias{split.elispot}
\title{\code{\link[=split.elispot]{split.elispot()}}}
\usage{
\method{split}{elispot}(x, f, drop = TRUE, ...)
}
\arguments{
\item{x}{an \linkS4class{elispot}}

\item{f}{..}

\item{drop}{\link[base]{logical} scalar, must be \code{FALSE}}
}
\description{
\code{\link[=split.elispot]{split.elispot()}}
}
\keyword{internal}
