% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_elispot.R
\name{trim_elispot}
\alias{trim_elispot}
\title{Trim \code{'elispot'} Object}
\usage{
trim_elispot(x)
}
\arguments{
\item{x}{an \code{elispot} object}
}
\value{
Function \link{trim_elispot} returns an \code{elispot} object.
}
\description{
Trim \code{'elispot'} Object
}
\details{
Both treatment \verb{$x1} and control \verb{$x0} must contain more than 1 measurement
}
